Split kv.h (#1179)
