+++ samples/apps/smallbank/fbs/tests/small_bank_serdes_bench.cpp
@@ -3,7 +3,6 @@
 #define PICOBENCH_IMPLEMENT
 
 #include "consensus/pbft/pbft_requests.h"
-#include "kv/kv.h"
 #include "node/encryptor.h"
 #include "node/history.h"
 #include "node/rpc/json_rpc.h"
@@ -67,7 +66,7 @@ static std::vector<uint8_t> kv_serialized_data(std::vector<uint8_t>& data)
 
   auto& map0 = kv_store.create<pbft::RequestsMap>("map0");
 
-  ccf::Store::Tx tx(kv_store.next_version());
+  ccf::Tx tx(kv_store.next_version());
   auto tx0 = tx.get_view(map0);
 
   tx0->put(0, {0, {}, data, {}});
