+++ src/apps/logging/logging_schema.h
@@ -3,7 +3,7 @@
 #pragma once
 #include "ds/json.h"
 
-namespace ccf
+namespace loggingapp
 {
   // Private record/get
   // Explicit target structs, macro-generated parsers + schema
