+++ src/consensus/pbft/libbyz/ledger_writer.h
@@ -6,7 +6,7 @@
 #include "consensus/pbft/pbft_pre_prepares.h"
 #include "consensus/pbft/pbft_requests.h"
 #include "consensus/pbft/pbft_types.h"
-#include "kv/kv.h"
+#include "kv/kv_types.h"
 #include "ledger.h"
 #include "new_view.h"
 #include "node/signatures.h"
@@ -30,6 +30,6 @@ public:
   virtual ~LedgerWriter() = default;
   kv::Version write_pre_prepare(Pre_prepare* pp);
   kv::Version write_pre_prepare(Pre_prepare* pp, View view);
-  kv::Version write_pre_prepare(ccf::Store::Tx& tx, Pre_prepare* pp);
+  kv::Version write_pre_prepare(ccf::Tx& tx, Pre_prepare* pp);
   void write_new_view(New_view* nv);
 };
