+++ src/enclave/rpc_handler.h
@@ -39,9 +39,7 @@ namespace enclave
     virtual ProcessPbftResp process_pbft(
       std::shared_ptr<enclave::RpcContext> ctx) = 0;
     virtual ProcessPbftResp process_pbft(
-      std::shared_ptr<enclave::RpcContext>,
-      ccf::Store::Tx& tx,
-      bool playback) = 0;
+      std::shared_ptr<enclave::RpcContext>, ccf::Tx& tx, bool playback) = 0;
     virtual crypto::Sha256Hash get_merkle_root() = 0;
     virtual void update_merkle_tree() = 0;
   };
