+++ src/kv/test/kv_test.cpp
@@ -1,30 +1,28 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the Apache 2.0 License.
-#define DOCTEST_CONFIG_IMPLEMENT_WITH_MAIN
-#include "../kv.h"
-#include "../kv_serialiser.h"
-#include "../node/entities.h"
-#include "../node/history.h"
 #include "ds/logger.h"
 #include "enclave/app_interface.h"
+#include "kv/kv.h"
+#include "kv/kv_serialiser.h"
 #include "node/encryptor.h"
+#include "node/entities.h"
+#include "node/history.h"
 
+#define DOCTEST_CONFIG_IMPLEMENT_WITH_MAIN
 #include <doctest/doctest.h>
 #include <msgpack/msgpack.hpp>
 #include <string>
 #include <vector>
 
-using namespace ccf;
-
 TEST_CASE("Map creation")
 {
-  Store kv_store;
+  ccf::Store kv_store;
   auto& map = kv_store.create<std::string, std::string>("map");
 
   INFO("Get a map that does not exist");
   {
     // Macros can't handle commas, so we need a single named template argument
-    using StringString = Store::Map<std::string, std::string>;
+    using StringString = ccf::Store::Map<std::string, std::string>;
     REQUIRE(kv_store.get<StringString>("invalid_map") == nullptr);
   }
 
@@ -36,30 +34,30 @@ TEST_CASE("Map creation")
 
   INFO("Can't create map that already exists");
   {
-    using StringString = Store::Map<std::string, std::string>;
+    using StringString = ccf::Store::Map<std::string, std::string>;
     REQUIRE_THROWS_AS(kv_store.create<StringString>("map"), std::logic_error);
   }
 
   INFO("Can't get a map with the wrong type");
   {
-    using IntInt = Store::Map<int, int>;
+    using IntInt = ccf::Store::Map<int, int>;
     REQUIRE(kv_store.get<IntInt>("map") == nullptr);
-    using IntString = Store::Map<int, std::string>;
+    using IntString = ccf::Store::Map<int, std::string>;
     REQUIRE(kv_store.get<IntString>("map") == nullptr);
-    using StringInt = Store::Map<std::string, int>;
+    using StringInt = ccf::Store::Map<std::string, int>;
     REQUIRE(kv_store.get<StringInt>("map") == nullptr);
   }
 
   INFO("Can create a map with a previously invalid name");
   {
-    using StringString = Store::Map<std::string, std::string>;
+    using StringString = ccf::Store::Map<std::string, std::string>;
     CHECK_NOTHROW(kv_store.create<StringString>("version"));
   }
 }
 
 TEST_CASE("Reads/writes and deletions")
 {
-  Store kv_store;
+  ccf::Store kv_store;
   auto& map = kv_store.create<std::string, std::string>(
     "map", kv::SecurityDomain::PUBLIC);
 
@@ -69,14 +67,14 @@ TEST_CASE("Reads/writes and deletions")
 
   INFO("Start empty transaction");
   {
-    Store::Tx tx;
+    ccf::Tx tx;
     REQUIRE(tx.commit() == kv::CommitSuccess::OK);
     REQUIRE_THROWS_AS(tx.commit(), std::logic_error);
   }
 
   INFO("Read own writes");
   {
-    Store::Tx tx;
+    ccf::Tx tx;
     auto view = tx.get_view(map);
     auto v = view->get(k);
     REQUIRE(!v.has_value());
@@ -89,7 +87,7 @@ TEST_CASE("Reads/writes and deletions")
 
   INFO("Read previous writes");
   {
-    Store::Tx tx;
+    ccf::Tx tx;
     auto view = tx.get_view(map);
     auto v = view->get(k);
     REQUIRE(v.has_value());
@@ -99,8 +97,8 @@ TEST_CASE("Reads/writes and deletions")
 
   INFO("Remove keys");
   {
-    Store::Tx tx;
-    Store::Tx tx2;
+    ccf::Tx tx;
+    ccf::Tx tx2;
     auto view = tx.get_view(map);
     view->put(k, v1);
 
@@ -117,9 +115,9 @@ TEST_CASE("Reads/writes and deletions")
 
   INFO("Remove key that was deleted from state");
   {
-    Store::Tx tx;
-    Store::Tx tx2;
-    Store::Tx tx3;
+    ccf::Tx tx;
+    ccf::Tx tx2;
+    ccf::Tx tx3;
     auto view = tx.get_view(map);
     view->put(k, v1);
     auto va = view->get_globally_committed(k);
@@ -136,7 +134,7 @@ TEST_CASE("Reads/writes and deletions")
 
   INFO("Test early temination of KV foreach");
   {
-    Store::Tx tx;
+    ccf::Tx tx;
     auto view = tx.get_view(map);
     view->put("key1", "value1");
     view->put("key2", "value2");
@@ -151,7 +149,7 @@ TEST_CASE("Reads/writes and deletions")
 
 TEST_CASE("Rollback and compact")
 {
-  Store kv_store;
+  ccf::Store kv_store;
   auto& map = kv_store.create<std::string, std::string>(
     "map", kv::SecurityDomain::PUBLIC);
 
@@ -160,8 +158,8 @@ TEST_CASE("Rollback and compact")
 
   INFO("Do not read transactions that have been rolled back");
   {
-    Store::Tx tx;
-    Store::Tx tx2;
+    ccf::Tx tx;
+    ccf::Tx tx2;
     auto view = tx.get_view(map);
     view->put(k, v1);
     REQUIRE(tx.commit() == kv::CommitSuccess::OK);
@@ -175,8 +173,8 @@ TEST_CASE("Rollback and compact")
 
   INFO("Read committed key");
   {
-    Store::Tx tx;
-    Store::Tx tx2;
+    ccf::Tx tx;
+    ccf::Tx tx2;
     auto view = tx.get_view(map);
     view->put(k, v1);
     REQUIRE(tx.commit() == kv::CommitSuccess::OK);
@@ -190,8 +188,8 @@ TEST_CASE("Rollback and compact")
 
   INFO("Read deleted committed key");
   {
-    Store::Tx tx;
-    Store::Tx tx2;
+    ccf::Tx tx;
+    ccf::Tx tx2;
     auto view = tx.get_view(map);
     REQUIRE(view->remove(k));
     REQUIRE(tx.commit() == kv::CommitSuccess::OK);
@@ -205,7 +203,7 @@ TEST_CASE("Rollback and compact")
 
 TEST_CASE("Clear entire store")
 {
-  Store kv_store;
+  ccf::Store kv_store;
   auto& map1 = kv_store.create<std::string, std::string>(
     "map1", kv::SecurityDomain::PUBLIC);
   auto& map2 = kv_store.create<std::string, std::string>(
@@ -213,8 +211,8 @@ TEST_CASE("Clear entire store")
 
   INFO("Commit a transaction over two maps");
   {
-    Store::Tx tx;
-    Store::Tx tx2;
+    ccf::Tx tx;
+    ccf::Tx tx2;
     auto [view1, view2] = tx.get_view(map1, map2);
     view1->put("key1", "value1");
     view2->put("key2", "value2");
@@ -227,8 +225,8 @@ TEST_CASE("Clear entire store")
 
   INFO("Clear the entire store and make sure it is empty");
   {
-    Store::Tx tx;
-    Store::Tx tx2;
+    ccf::Tx tx;
+    ccf::Tx tx2;
     auto [view1, view2] = tx.get_view(map1, map2);
 
     kv_store.clear();
@@ -244,8 +242,8 @@ TEST_CASE("Clear entire store")
 
 TEST_CASE("Local commit hooks")
 {
-  using State = Store::Map<std::string, std::string>::State;
-  using Write = Store::Map<std::string, std::string>::Write;
+  using State = ccf::Store::Map<std::string, std::string>::State;
+  using Write = ccf::Store::Map<std::string, std::string>::Write;
   std::vector<Write> local_writes;
   std::vector<Write> global_writes;
 
@@ -256,13 +254,13 @@ TEST_CASE("Local commit hooks")
     global_writes.push_back(w);
   };
 
-  Store kv_store;
+  ccf::Store kv_store;
   auto& map = kv_store.create<std::string, std::string>(
     "map", kv::SecurityDomain::PUBLIC, local_hook, global_hook);
 
   INFO("Write with hooks");
   {
-    Store::Tx tx;
+    ccf::Tx tx;
     auto view = tx.get_view(map);
     view->put("key", "value1");
     REQUIRE(tx.commit() == kv::CommitSuccess::OK);
@@ -279,7 +277,7 @@ TEST_CASE("Local commit hooks")
     map.unset_local_hook();
     map.unset_global_hook();
 
-    Store::Tx tx;
+    ccf::Tx tx;
     auto view = tx.get_view(map);
     view->put("key", "value2");
     REQUIRE(tx.commit() == kv::CommitSuccess::OK);
@@ -293,7 +291,7 @@ TEST_CASE("Local commit hooks")
     map.set_local_hook(local_hook);
     map.set_global_hook(global_hook);
 
-    Store::Tx tx;
+    ccf::Tx tx;
     auto view = tx.get_view(map);
     view->put("key", "value3");
     REQUIRE(tx.commit() == kv::CommitSuccess::OK);
@@ -308,8 +306,8 @@ TEST_CASE("Local commit hooks")
 
 TEST_CASE("Global commit hooks")
 {
-  using State = Store::Map<std::string, std::string>::State;
-  using Write = Store::Map<std::string, std::string>::Write;
+  using State = ccf::Store::Map<std::string, std::string>::State;
+  using Write = ccf::Store::Map<std::string, std::string>::Write;
   struct GlobalHookInput
   {
     kv::Version version;
@@ -322,7 +320,7 @@ TEST_CASE("Global commit hooks")
     global_writes.emplace_back(GlobalHookInput({v, w}));
   };
 
-  Store kv_store;
+  ccf::Store kv_store;
   auto& map_with_hook = kv_store.create<std::string, std::string>(
     "map_with_hook", kv::SecurityDomain::PUBLIC, nullptr, global_hook);
   auto& map_no_hook = kv_store.create<std::string, std::string>(
@@ -337,7 +335,7 @@ TEST_CASE("Global commit hooks")
 
   INFO("Compact one transaction");
   {
-    Store::Tx tx1;
+    ccf::Tx tx1;
     auto view_hook = tx1.get_view(map_with_hook);
     view_hook->put("key1", "value1");
     REQUIRE(tx1.commit() == kv::CommitSuccess::OK);
@@ -354,7 +352,7 @@ TEST_CASE("Global commit hooks")
 
   INFO("Compact beyond the last map version");
   {
-    Store::Tx tx1, tx2, tx3;
+    ccf::Tx tx1, tx2, tx3;
     auto view_hook = tx1.get_view(map_with_hook);
     view_hook->put("key1", "value1");
     REQUIRE(tx1.commit() == kv::CommitSuccess::OK);
@@ -385,7 +383,7 @@ TEST_CASE("Global commit hooks")
 
   INFO("Compact in between two map versions");
   {
-    Store::Tx tx1, tx2, tx3;
+    ccf::Tx tx1, tx2, tx3;
     auto view_hook = tx1.get_view(map_with_hook);
     view_hook->put("key1", "value1");
     REQUIRE(tx1.commit() == kv::CommitSuccess::OK);
@@ -414,7 +412,7 @@ TEST_CASE("Global commit hooks")
 
   INFO("Compact twice");
   {
-    Store::Tx tx1, tx2;
+    ccf::Tx tx1, tx2;
     auto view_hook = tx1.get_view(map_with_hook);
     view_hook->put("key1", "value1");
     REQUIRE(tx1.commit() == kv::CommitSuccess::OK);
@@ -441,20 +439,20 @@ TEST_CASE("Global commit hooks")
 TEST_CASE("Clone schema")
 {
   auto encryptor = std::make_shared<ccf::NullTxEncryptor>();
-  Store store;
+  ccf::Store store;
   store.set_encryptor(encryptor);
 
   auto& public_map =
     store.create<size_t, std::string>("public", kv::SecurityDomain::PUBLIC);
   auto& private_map = store.create<size_t, std::string>("private");
-  Store::Tx tx1(store.next_version());
+  ccf::Tx tx1(store.next_version());
   auto [view1, view2] = tx1.get_view(public_map, private_map);
   view1->put(42, "aardvark");
   view2->put(14, "alligator");
   auto [success, reqid, data] = tx1.commit_reserved();
   REQUIRE(success == kv::CommitSuccess::OK);
 
-  Store clone;
+  ccf::Store clone;
   clone.clone_schema(store);
   clone.set_encryptor(encryptor);
 
@@ -463,7 +461,7 @@ TEST_CASE("Clone schema")
 
 TEST_CASE("Deserialise return status")
 {
-  Store store;
+  ccf::Store store;
 
   auto& signatures = store.create<ccf::Signatures>(
     ccf::Tables::SIGNATURES, kv::SecurityDomain::PUBLIC);
@@ -478,7 +476,7 @@ TEST_CASE("Deserialise return status")
   store.set_history(history);
 
   {
-    Store::Tx tx(store.next_version());
+    ccf::Tx tx(store.next_version());
     auto data_view = tx.get_view(data);
     data_view->put(42, 42);
     auto [success, reqid, data] = tx.commit_reserved();
@@ -488,7 +486,7 @@ TEST_CASE("Deserialise return status")
   }
 
   {
-    Store::Tx tx(store.next_version());
+    ccf::Tx tx(store.next_version());
     auto sig_view = tx.get_view(signatures);
     ccf::Signature sigv(0, 2);
     sig_view->put(0, sigv);
@@ -500,7 +498,7 @@ TEST_CASE("Deserialise return status")
 
   INFO("Signature transactions with additional contents should fail");
   {
-    Store::Tx tx(store.next_version());
+    ccf::Tx tx(store.next_version());
     auto [sig_view, data_view] = tx.get_view(signatures, data);
     ccf::Signature sigv(0, 2);
     sig_view->put(0, sigv);
@@ -515,35 +513,35 @@ TEST_CASE("Deserialise return status")
 TEST_CASE("map swap between stores")
 {
   auto encryptor = std::make_shared<ccf::NullTxEncryptor>();
-  Store s1;
+  ccf::Store s1;
   s1.set_encryptor(encryptor);
 
   auto& d1 = s1.create<size_t, size_t>("data");
   auto& pd1 =
     s1.create<size_t, size_t>("public_data", kv::SecurityDomain::PUBLIC);
 
-  Store s2;
+  ccf::Store s2;
   s2.set_encryptor(encryptor);
   auto& d2 = s2.create<size_t, size_t>("data");
   auto& pd2 =
     s2.create<size_t, size_t>("public_data", kv::SecurityDomain::PUBLIC);
 
   {
-    Store::Tx tx;
+    ccf::Tx tx;
     auto v = tx.get_view(d1);
     v->put(42, 42);
     tx.commit();
   }
 
   {
-    Store::Tx tx;
+    ccf::Tx tx;
     auto v = tx.get_view(pd1);
     v->put(14, 14);
     tx.commit();
   }
 
   {
-    Store::Tx tx;
+    ccf::Tx tx;
     auto v = tx.get_view(d2);
     v->put(41, 41);
     tx.commit();
@@ -552,7 +550,7 @@ TEST_CASE("map swap between stores")
   s2.swap_private_maps(s1);
 
   {
-    Store::Tx tx;
+    ccf::Tx tx;
     auto v = tx.get_view(d1);
     auto val = v->get(41);
     REQUIRE_FALSE(v->get(42).has_value());
@@ -561,7 +559,7 @@ TEST_CASE("map swap between stores")
   }
 
   {
-    Store::Tx tx;
+    ccf::Tx tx;
     auto v = tx.get_view(pd1);
     auto val = v->get(14);
     REQUIRE(val.has_value());
@@ -569,7 +567,7 @@ TEST_CASE("map swap between stores")
   }
 
   {
-    Store::Tx tx;
+    ccf::Tx tx;
     auto v = tx.get_view(d2);
     auto val = v->get(42);
     REQUIRE_FALSE(v->get(41).has_value());
@@ -578,7 +576,7 @@ TEST_CASE("map swap between stores")
   }
 
   {
-    Store::Tx tx;
+    ccf::Tx tx;
     auto v = tx.get_view(pd2);
     REQUIRE_FALSE(v->get(14).has_value());
   }
@@ -587,10 +585,10 @@ TEST_CASE("map swap between stores")
 TEST_CASE("invalid map swaps")
 {
   {
-    Store s1;
+    ccf::Store s1;
     s1.create<size_t, size_t>("one");
 
-    Store s2;
+    ccf::Store s2;
     s2.create<size_t, size_t>("one");
     s2.create<size_t, size_t>("two");
 
@@ -600,11 +598,11 @@ TEST_CASE("invalid map swaps")
   }
 
   {
-    Store s1;
+    ccf::Store s1;
     s1.create<size_t, size_t>("one");
     s1.create<size_t, size_t>("two");
 
-    Store s2;
+    ccf::Store s2;
     s2.create<size_t, size_t>("one");
 
     REQUIRE_THROWS_WITH(
@@ -616,13 +614,13 @@ TEST_CASE("invalid map swaps")
 TEST_CASE("private recovery map swap")
 {
   auto encryptor = std::make_shared<ccf::NullTxEncryptor>();
-  Store s1;
+  ccf::Store s1;
   s1.set_encryptor(encryptor);
   auto& priv1 = s1.create<size_t, size_t>("private");
   auto& pub1 =
     s1.create<size_t, std::string>("public", kv::SecurityDomain::PUBLIC);
 
-  Store s2;
+  ccf::Store s2;
   s2.set_encryptor(encryptor);
   auto& priv2 = s2.create<size_t, size_t>("private");
   auto& pub2 =
@@ -632,27 +630,27 @@ TEST_CASE("private recovery map swap")
   // We compact twice, deliberately. A public KV during recovery
   // would have compacted some number of times.
   {
-    Store::Tx tx;
+    ccf::Tx tx;
     auto v = tx.get_view(pub1);
     v->put(42, "42");
     tx.commit();
   }
   {
-    Store::Tx tx;
+    ccf::Tx tx;
     auto v = tx.get_view(pub1);
     v->put(42, "43");
     tx.commit();
   }
   s1.compact(s1.current_version());
   {
-    Store::Tx tx;
+    ccf::Tx tx;
     auto v = tx.get_view(pub1);
     v->put(44, "44");
     tx.commit();
   }
   s1.compact(s1.current_version());
   {
-    Store::Tx tx;
+    ccf::Tx tx;
     auto v = tx.get_view(pub1);
     v->put(45, "45");
     tx.commit();
@@ -660,27 +658,28 @@ TEST_CASE("private recovery map swap")
 
   INFO("Populate s2 with private entries");
   // We compact only once, at a lower index than we did for the public
-  // KV, which is what we expect during recovery of the private KV.
+  // KV, which is what we expect during recovery of the private KV. We do expect
+  // that the _entire_ private state is compacted
   {
-    Store::Tx tx;
+    ccf::Tx tx;
     auto v = tx.get_view(priv2);
     v->put(12, 12);
     tx.commit();
   }
-  s2.compact(s2.current_version());
   {
-    Store::Tx tx;
+    ccf::Tx tx;
     auto v = tx.get_view(priv2);
     v->put(13, 13);
     tx.commit();
   }
+  s2.compact(s2.current_version());
 
   INFO("Swap in private maps");
   s1.swap_private_maps(s2);
 
   INFO("Check state looks as expected in s1");
   {
-    Store::Tx tx;
+    ccf::Tx tx;
     auto [priv, pub] = tx.get_view(priv1, pub1);
     {
       auto val = pub->get(42);
@@ -710,27 +709,26 @@ TEST_CASE("private recovery map swap")
 
   INFO("Check committed state looks as expected in s1");
   {
-    Store::Tx tx;
-    tx.set_read_committed();
+    ccf::Tx tx;
     auto [priv, pub] = tx.get_view(priv1, pub1);
     {
-      auto val = pub->get(42);
+      auto val = pub->get_globally_committed(42);
       REQUIRE(val.has_value());
       REQUIRE(val.value() == "43");
 
-      val = pub->get(44);
+      val = pub->get_globally_committed(44);
       REQUIRE(val.has_value());
       REQUIRE(val.value() == "44");
 
-      val = pub->get(45);
+      val = pub->get_globally_committed(45);
       REQUIRE_FALSE(val.has_value());
     }
     {
-      auto val = priv->get(12);
+      auto val = priv->get_globally_committed(12);
       REQUIRE(val.has_value());
       REQUIRE(val.value() == 12);
 
-      val = priv->get(13);
+      val = priv->get_globally_committed(13);
       REQUIRE(val.has_value());
       REQUIRE(val.value() == 13);
 
@@ -745,11 +743,11 @@ TEST_CASE("private recovery map swap")
 
 TEST_CASE("Conflict resolution")
 {
-  Store kv_store;
+  ccf::Store kv_store;
   auto& map = kv_store.create<std::string, std::string>(
     "map", kv::SecurityDomain::PUBLIC);
 
-  auto try_write = [&](Store::Tx& tx, const std::string& s) {
+  auto try_write = [&](ccf::Tx& tx, const std::string& s) {
     auto view = tx.get_view(map);
 
     // Introduce read-dependency
@@ -762,7 +760,7 @@ TEST_CASE("Conflict resolution")
   auto confirm_state = [&](
                          const std::vector<std::string>& present,
                          const std::vector<std::string>& missing) {
-    Store::Tx tx;
+    ccf::Tx tx;
     auto view = tx.get_view(map);
 
     for (const auto& s : present)
@@ -780,8 +778,8 @@ TEST_CASE("Conflict resolution")
   };
 
   // Simulate parallel execution by interleaving tx steps
-  Store::Tx tx1;
-  Store::Tx tx2;
+  ccf::Tx tx1;
+  ccf::Tx tx2;
 
   // First transaction tries to write a value, depending on initial version
   try_write(tx1, "bar");
