+++ src/node/entities.h
@@ -69,4 +69,5 @@ namespace ccf
   using StoreSerialiser = kv::KvStoreSerialiser;
   using StoreDeserialiser = kv::KvStoreDeserialiser;
   using Store = kv::Store<StoreSerialiser, StoreDeserialiser>;
+  using Tx = kv::Tx<StoreSerialiser, StoreDeserialiser>;
 }
