+++ src/node/rpc/node_interface.h
@@ -12,22 +12,22 @@ namespace ccf
   {
   public:
     virtual ~AbstractNodeState() {}
-    virtual bool accept_recovery(Store::Tx& tx) = 0;
-    virtual bool open_network(Store::Tx& tx) = 0;
-    virtual bool rekey_ledger(Store::Tx& tx) = 0;
+    virtual bool accept_recovery(ccf::Tx& tx) = 0;
+    virtual bool open_network(ccf::Tx& tx) = 0;
+    virtual bool rekey_ledger(ccf::Tx& tx) = 0;
     virtual bool is_part_of_public_network() const = 0;
     virtual bool is_primary() const = 0;
     virtual bool is_reading_public_ledger() const = 0;
     virtual bool is_reading_private_ledger() const = 0;
     virtual bool is_part_of_network() const = 0;
     virtual void node_quotes(
-      Store::Tx& tx,
+      ccf::Tx& tx,
       GetQuotes::Out& result,
       const std::optional<std::set<NodeId>>& filter = std::nullopt) = 0;
     virtual NodeId get_node_id() const = 0;
 
-    virtual bool split_ledger_secrets(Store::Tx& tx) = 0;
+    virtual bool split_ledger_secrets(ccf::Tx& tx) = 0;
     virtual bool restore_ledger_secrets(
-      Store::Tx& tx, const std::vector<SecretSharing::Share>& shares) = 0;
+      ccf::Tx& tx, const std::vector<SecretSharing::Share>& shares) = 0;
   };
 }
