+++ src/node/rpc/test/member_voting_test.cpp
@@ -195,7 +195,7 @@ DOCTEST_TEST_CASE("Member query/read")
 {
   // initialize the network state
   NetworkTables network;
-  Store::Tx gen_tx;
+  ccf::Tx gen_tx;
   GenesisGenerator gen(network, gen_tx);
   gen.init_values();
   gen.create_service({});
@@ -211,7 +211,7 @@ DOCTEST_TEST_CASE("Member query/read")
   // put value to read
   constexpr auto key = 123;
   constexpr auto value = 456;
-  Store::Tx tx;
+  ccf::Tx tx;
   tx.get_view(network.values)->put(key, value);
   DOCTEST_CHECK(tx.commit() == kv::CommitSuccess::OK);
 
@@ -223,7 +223,7 @@ DOCTEST_TEST_CASE("Member query/read")
   DOCTEST_SUBCASE("Query: bytecode/script allowed access")
   {
     // set member ACL so that the VALUES table is accessible
-    Store::Tx tx;
+    ccf::Tx tx;
     tx.get_view(network.whitelists)
       ->put(WlIds::MEMBER_CAN_READ, {Tables::VALUES});
     DOCTEST_CHECK(tx.commit() == kv::CommitSuccess::OK);
@@ -242,7 +242,7 @@ DOCTEST_TEST_CASE("Member query/read")
   DOCTEST_SUBCASE("Query: table not in ACL")
   {
     // set member ACL so that no table is accessible
-    Store::Tx tx;
+    ccf::Tx tx;
     tx.get_view(network.whitelists)->put(WlIds::MEMBER_CAN_READ, {});
     DOCTEST_CHECK(tx.commit() == kv::CommitSuccess::OK);
 
@@ -254,7 +254,7 @@ DOCTEST_TEST_CASE("Member query/read")
 
   DOCTEST_SUBCASE("Read: allowed access, key exists")
   {
-    Store::Tx tx;
+    ccf::Tx tx;
     tx.get_view(network.whitelists)
       ->put(WlIds::MEMBER_CAN_READ, {Tables::VALUES});
     DOCTEST_CHECK(tx.commit() == kv::CommitSuccess::OK);
@@ -269,7 +269,7 @@ DOCTEST_TEST_CASE("Member query/read")
   DOCTEST_SUBCASE("Read: allowed access, key doesn't exist")
   {
     constexpr auto wrong_key = 321;
-    Store::Tx tx;
+    ccf::Tx tx;
     tx.get_view(network.whitelists)
       ->put(WlIds::MEMBER_CAN_READ, {Tables::VALUES});
     DOCTEST_CHECK(tx.commit() == kv::CommitSuccess::OK);
@@ -283,7 +283,7 @@ DOCTEST_TEST_CASE("Member query/read")
 
   DOCTEST_SUBCASE("Read: access not allowed")
   {
-    Store::Tx tx;
+    ccf::Tx tx;
     tx.get_view(network.whitelists)->put(WlIds::MEMBER_CAN_READ, {});
     DOCTEST_CHECK(tx.commit() == kv::CommitSuccess::OK);
 
@@ -299,7 +299,7 @@ DOCTEST_TEST_CASE("Proposer ballot")
 {
   NetworkTables network;
   network.tables->set_encryptor(encryptor);
-  Store::Tx gen_tx;
+  ccf::Tx gen_tx;
   GenesisGenerator gen(network, gen_tx);
   gen.init_values();
   gen.create_service({});
@@ -401,7 +401,7 @@ DOCTEST_TEST_CASE("Add new members until there are 7 then reject")
   constexpr auto max_members = 8;
   NetworkTables network;
   network.tables->set_encryptor(encryptor);
-  Store::Tx gen_tx;
+  ccf::Tx gen_tx;
   GenesisGenerator gen(network, gen_tx);
   gen.init_values();
   gen.create_service({});
@@ -549,7 +549,7 @@ DOCTEST_TEST_CASE("Add new members until there are 7 then reject")
       {
         // make sure that there is a signature in the signatures table since
         // ack's depend on that
-        Store::Tx tx;
+        ccf::Tx tx;
         auto sig_view = tx.get_view(network.signatures);
         Signature sig_value;
         sig_view->put(0, sig_value);
@@ -601,7 +601,7 @@ DOCTEST_TEST_CASE("Accept node")
 {
   NetworkTables network;
   network.tables->set_encryptor(encryptor);
-  Store::Tx gen_tx;
+  ccf::Tx gen_tx;
   GenesisGenerator gen(network, gen_tx);
   gen.init_values();
   gen.create_service({});
@@ -756,7 +756,7 @@ ProposalInfo test_raw_writes(
 
   // check values before
   {
-    Store::Tx tx;
+    ccf::Tx tx;
     auto next_member_id_r =
       tx.get_view(network.values)->get(ValueIds::NEXT_MEMBER_ID);
     DOCTEST_CHECK(next_member_id_r);
@@ -825,14 +825,14 @@ DOCTEST_TEST_CASE("Propose raw writes")
       const bool should_succeed = pro_votes > n_members / 2;
       NetworkTables network;
       network.tables->set_encryptor(encryptor);
-      Store::Tx gen_tx;
+      ccf::Tx gen_tx;
       GenesisGenerator gen(network, gen_tx);
       gen.init_values();
       gen.create_service({});
       StubNodeState node;
       nlohmann::json recovery_threshold = 4;
 
-      Store::Tx tx_before;
+      ccf::Tx tx_before;
       auto configuration = tx_before.get_view(network.config)->get(0);
       DOCTEST_REQUIRE_FALSE(configuration.has_value());
 
@@ -856,7 +856,7 @@ DOCTEST_TEST_CASE("Propose raw writes")
         continue;
 
       // check results
-      Store::Tx tx_after;
+      ccf::Tx tx_after;
       configuration = tx_after.get_view(network.config)->get(0);
       DOCTEST_CHECK(configuration.has_value());
       DOCTEST_CHECK(configuration->recovery_threshold == recovery_threshold);
@@ -878,7 +878,7 @@ DOCTEST_TEST_CASE("Propose raw writes")
         {
           NetworkTables network;
           network.tables->set_encryptor(encryptor);
-          Store::Tx gen_tx;
+          ccf::Tx gen_tx;
           GenesisGenerator gen(network, gen_tx);
           gen.init_values();
           gen.create_service({});
@@ -914,7 +914,7 @@ DOCTEST_TEST_CASE("Remove proposal")
 
   NetworkTables network;
   network.tables->set_encryptor(encryptor);
-  Store::Tx gen_tx;
+  ccf::Tx gen_tx;
   GenesisGenerator gen(network, gen_tx);
   gen.init_values();
   gen.create_service({});
@@ -936,7 +936,7 @@ DOCTEST_TEST_CASE("Remove proposal")
 
   // check that the proposal doesn't exist
   {
-    Store::Tx tx;
+    ccf::Tx tx;
     auto proposal = tx.get_view(network.proposals)->get(proposal_id);
     DOCTEST_CHECK(!proposal);
   }
@@ -953,7 +953,7 @@ DOCTEST_TEST_CASE("Remove proposal")
 
   // check that the proposal is there
   {
-    Store::Tx tx;
+    ccf::Tx tx;
     auto proposal = tx.get_view(network.proposals)->get(proposal_id);
     DOCTEST_CHECK(proposal);
     DOCTEST_CHECK(proposal->state == ProposalState::OPEN);
@@ -994,7 +994,7 @@ DOCTEST_TEST_CASE("Remove proposal")
 
     // check that the proposal is now withdrawn
     {
-      Store::Tx tx;
+      ccf::Tx tx;
       auto proposal = tx.get_view(network.proposals)->get(proposal_id);
       DOCTEST_CHECK(proposal.has_value());
       DOCTEST_CHECK(proposal->state == ProposalState::WITHDRAWN);
@@ -1006,7 +1006,7 @@ DOCTEST_TEST_CASE("Complete proposal after initial rejection")
 {
   NetworkTables network;
   network.tables->set_encryptor(encryptor);
-  Store::Tx gen_tx;
+  ccf::Tx gen_tx;
   GenesisGenerator gen(network, gen_tx);
   gen.init_values();
   gen.create_service({});
@@ -1022,7 +1022,7 @@ DOCTEST_TEST_CASE("Complete proposal after initial rejection")
     const auto propose =
       create_signed_request(Propose::In{proposal}, "propose", kp);
 
-    Store::Tx tx;
+    ccf::Tx tx;
     const auto r = parse_response_body<Propose::Out>(
       frontend_process(frontend, propose, member_certs[0]));
     DOCTEST_CHECK(r.state == ProposalState::OPEN);
@@ -1054,7 +1054,7 @@ DOCTEST_TEST_CASE("Complete proposal after initial rejection")
 
   {
     DOCTEST_INFO("Put value that makes vote agree");
-    Store::Tx tx;
+    ccf::Tx tx;
     tx.get_view(network.values)->put(123, 123);
     DOCTEST_CHECK(tx.commit() == kv::CommitSuccess::OK);
   }
@@ -1074,7 +1074,7 @@ DOCTEST_TEST_CASE("Vetoed proposal gets rejected")
 {
   NetworkTables network;
   network.tables->set_encryptor(encryptor);
-  Store::Tx gen_tx;
+  ccf::Tx gen_tx;
   GenesisGenerator gen(network, gen_tx);
   gen.init_values();
   gen.create_service({});
@@ -1126,7 +1126,7 @@ DOCTEST_TEST_CASE("Add user via proposed call")
 {
   NetworkTables network;
   network.tables->set_encryptor(encryptor);
-  Store::Tx gen_tx;
+  ccf::Tx gen_tx;
   GenesisGenerator gen(network, gen_tx);
   gen.init_values();
   gen.create_service({});
@@ -1153,7 +1153,7 @@ DOCTEST_TEST_CASE("Add user via proposed call")
   DOCTEST_CHECK(r.state == ProposalState::ACCEPTED);
   DOCTEST_CHECK(r.proposal_id == 0);
 
-  Store::Tx tx1;
+  ccf::Tx tx1;
   const auto uid = tx1.get_view(network.values)->get(ValueIds::NEXT_USER_ID);
   DOCTEST_CHECK(uid);
   DOCTEST_CHECK(*uid == 1);
@@ -1169,7 +1169,7 @@ DOCTEST_TEST_CASE("Passing members ballot with operator")
   // Operator votes, but is _not_ taken into consideration
   NetworkTables network;
   network.tables->set_encryptor(encryptor);
-  Store::Tx gen_tx;
+  ccf::Tx gen_tx;
   GenesisGenerator gen(network, gen_tx);
   gen.init_values();
   gen.create_service({});
@@ -1278,7 +1278,7 @@ DOCTEST_TEST_CASE("Passing operator vote")
   // and gets it through without member votes
   NetworkTables network;
   network.tables->set_encryptor(encryptor);
-  Store::Tx gen_tx;
+  ccf::Tx gen_tx;
   GenesisGenerator gen(network, gen_tx);
   gen.init_values();
   gen.create_service({});
@@ -1364,7 +1364,7 @@ DOCTEST_TEST_CASE("Members passing an operator vote")
   // A majority of members pass the vote
   NetworkTables network;
   network.tables->set_encryptor(encryptor);
-  Store::Tx gen_tx;
+  ccf::Tx gen_tx;
   GenesisGenerator gen(network, gen_tx);
   gen.init_values();
   gen.create_service({});
@@ -1478,7 +1478,7 @@ DOCTEST_TEST_CASE("User data")
 {
   NetworkTables network;
   network.tables->set_encryptor(encryptor);
-  Store::Tx gen_tx;
+  ccf::Tx gen_tx;
   GenesisGenerator gen(network, gen_tx);
   gen.init_values();
   gen.create_service({});
@@ -1570,7 +1570,7 @@ DOCTEST_TEST_CASE("Submit recovery shares")
 
   DOCTEST_INFO("Setup state");
   {
-    Store::Tx gen_tx;
+    ccf::Tx gen_tx;
 
     GenesisGenerator gen(network, gen_tx);
     gen.init_values();
@@ -1614,7 +1614,7 @@ DOCTEST_TEST_CASE("Submit recovery shares")
 
   DOCTEST_INFO("Change service state to waiting for recovery shares");
   {
-    Store::Tx tx;
+    ccf::Tx tx;
     GenesisGenerator g(network, tx);
 
     DOCTEST_REQUIRE(g.service_wait_for_shares());
@@ -1663,7 +1663,7 @@ DOCTEST_TEST_CASE("Maximum number of active members")
 
   DOCTEST_INFO("Service is opening");
   {
-    Store::Tx gen_tx;
+    ccf::Tx gen_tx;
     GenesisGenerator gen(network, gen_tx);
     gen.init_values();
     gen.create_service({});
@@ -1691,7 +1691,7 @@ DOCTEST_TEST_CASE("Maximum number of active members")
   {
     std::map<size_t, ccf::Cert> members;
 
-    Store::Tx gen_tx;
+    ccf::Tx gen_tx;
     GenesisGenerator gen(network, gen_tx);
     gen.init_values();
     gen.create_service({});
