+++ src/node/rpc/test/node_frontend_test.cpp
@@ -76,7 +76,7 @@ T parse_response_body(const TResponse& r)
 TEST_CASE("Add a node to an opening service")
 {
   NetworkState network;
-  Store::Tx gen_tx;
+  ccf::Tx gen_tx;
   GenesisGenerator gen(network, gen_tx);
   gen.init_values();
 
@@ -143,7 +143,7 @@ TEST_CASE("Add a node to an opening service")
     CHECK(response.node_status == NodeStatus::TRUSTED);
     CHECK(response.public_only == false);
 
-    Store::Tx tx;
+    ccf::Tx tx;
     const NodeId node_id = response.node_id;
     auto nodes_view = tx.get_view(network.nodes);
     auto node_info = nodes_view->get(node_id);
@@ -196,7 +196,7 @@ TEST_CASE("Add a node to an opening service")
 TEST_CASE("Add a node to an open service")
 {
   NetworkState network;
-  Store::Tx gen_tx;
+  ccf::Tx gen_tx;
   GenesisGenerator gen(network, gen_tx);
   gen.init_values();
 
@@ -221,7 +221,7 @@ TEST_CASE("Add a node to an open service")
   Cert caller = v->der_cert_data();
 
   std::optional<NodeInfo> node_info;
-  Store::Tx tx;
+  ccf::Tx tx;
 
   JoinNetworkNodeToNode::In join_input;
